+++ app/models/GameProps.tsx
