+++ app/pages/EditPage.tsx
@@ -1,17 +1,20 @@
-import { StyleSheet, ImageBackground} from 'react-native';
+// Import librairies
+import { StyleSheet, ImageBackground } from 'react-native';
 import { useLocalSearchParams } from "expo-router";
 import { SQLiteProvider } from 'expo-sqlite';
-import Game from '../Component/Game';
 
-export default function GamePage() {
+// Import components
+import Edit from '../components/cores/Edit';
+
+export default function EditPage() {
   const { plateauId } = useLocalSearchParams();
 
   return (
     <SQLiteProvider databaseName="mydatabasetest3.db" assetSource={{ assetId: require('../../assets/db/mydatabasetest3.db') }}>
       <ImageBackground source={require('../../assets/bg/bg_1.png')} style={styles.background}>
 
-        <Game plateauId={Array.isArray(plateauId) ? plateauId[0] : plateauId} />
-        
+        <Edit plateauId={Array.isArray(plateauId) ? plateauId[0] : plateauId}/>
+
       </ImageBackground>
     </SQLiteProvider>
   );
@@ -20,7 +23,8 @@ export default function GamePage() {
 const styles = StyleSheet.create({
   background: {
     flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
   }
 });
+
+
+
