+++ app/pages/GamePage.tsx
@@ -1,19 +1,20 @@
-import { View, StyleSheet, ImageBackground } from 'react-native';
+// Import librairies
+import { StyleSheet, ImageBackground} from 'react-native';
 import { useLocalSearchParams } from "expo-router";
 import { SQLiteProvider } from 'expo-sqlite';
 
-// Import Component
-import Edit from '../Component/Core/Edit';
+// Import components
+import Game from '../components/cores/Game';
 
-export default function EditPage() {
+export default function GamePage() {
   const { plateauId } = useLocalSearchParams();
 
   return (
     <SQLiteProvider databaseName="mydatabasetest3.db" assetSource={{ assetId: require('../../assets/db/mydatabasetest3.db') }}>
       <ImageBackground source={require('../../assets/bg/bg_1.png')} style={styles.background}>
 
-        <Edit plateauId={Array.isArray(plateauId) ? plateauId[0] : plateauId}/>
-
+        <Game plateauId={Array.isArray(plateauId) ? plateauId[0] : plateauId} />
+        
       </ImageBackground>
     </SQLiteProvider>
   );
@@ -22,8 +23,7 @@ export default function EditPage() {
 const styles = StyleSheet.create({
   background: {
     flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
   }
 });
-
-
-
